Move intro to camera
