+++ js/modules/HomeScene/Camera.js
@@ -1,5 +1,6 @@
 import * as THREE from "three";
 import { OrbitControls } from "three/examples/jsm/controls/OrbitControls.js";
+import gsap from "gsap";
 
 import HomeScene from "./HomeScene";
 
@@ -16,6 +17,8 @@ export default class Camera {
 
         this.setInstance();
         this.setControls();
+
+        this.intro = this.intro.bind(this);
     }
 
     setInstance() {
@@ -59,8 +62,33 @@ export default class Camera {
             // Update controls for rotating around root mesh
             controls.update();
         } else {
-            // Force camera to look at root mesh
+            // Force camera to look at center
             camera.lookAt(new THREE.Vector3());
         }
+    }
+
+    intro() {
+        const { instance: camera } = this;
+        // Intro animation
+        const tl = gsap.timeline({
+            paused: true,
+        });
+        tl.fromTo(
+            camera.position,
+            {
+                x: 2,
+                y: 0,
+                z: 2,
+            },
+            {
+                duration: 2,
+                ease: "power2.inOut",
+                x: 0,
+                y: 4,
+                z: 2,
+            },
+            0
+        );
+        tl.play();
     }
 }
